requirements.txt
# Sysrupt OT Range - Python dependencies

# Core infrastructure
redis>=5.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Industrial protocols
pymodbus>=3.5.0
cpppo>=4.3.0
python-snap7>=1.3
c104>=2.0.0
BAC0>=2023.6
opcua>=0.98.13
asyncua>=1.0.0

# Numerics / physics sim
numpy>=1.24.0
scipy>=1.11.0

# Display / imaging
luma.lcd>=2.13.0
luma.core>=2.4.0
Pillow>=10.0.0

# Hardware I/O
smbus2>=0.4.3
gpiod>=2.0.0
spidev>=3.6

# Web SCADA HMI
flask>=3.0.0
flask-socketio>=5.3.0
gunicorn>=21.2.0

# Network / IDS
scapy>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
